Move PaginatedIterable tests to bases/
